requirements.txt
codecov
coreapi
Django>=4.2
djangorestframework
django-cors-headers
django-crontab
django-storages
drf-yasg
flake8
importlib-metadata
packaging
pillow
psycopg2
pur
setuptools
six
tblib
tox
user-agents
XlsxWriter
xmltodict
